requirements.txt
# A 股自动监控系统依赖

# 数据采集
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
akshare>=1.10.0
baostock>=0.8.8
tushare>=1.2.80

# 数据处理
pandas>=1.5.0
numpy>=1.23.0

# NLP 情感分析
snownlp>=0.12.3
jieba>=0.42.1

# 技术分析
pandas-ta>=0.3.14b
# TA-Lib 需要手动安装：https://github.com/cgohlke/talib-build/releases
# 或者只用 pandas-ta 也可以满足需求

# 数据存储
sqlalchemy>=2.0.0

# 任务调度
apscheduler>=3.9.0

# 配置文件
pyyaml>=6.0

# 可视化与通知
rich>=13.0.0
matplotlib>=3.6.0

# 日志
loguru>=0.6.0

# 测试
pytest>=7.2.0
